+++ app/api/v1/__init__.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter
 from app.api.v1.endpoints.auth import router as auth_router
 from app.api.v1.endpoints.verification import router as verification_router
+from app.api.v1.endpoints.user import router as user_router
 
 router = APIRouter()
 
@@ -10,3 +11,4 @@ async def main_page():
 
 router.include_router(auth_router)
 router.include_router(verification_router)
+router.include_router(user_router)