+++ app/api/v1/endpoints/verification.py
@@ -1,10 +1,11 @@
 from fastapi import APIRouter, Depends, status, Request
-from sqlalchemy.ext.asyncio import AsyncSession
-from app.dependencies import get_verification_service, get_current_user
+from app import get_verification_service, get_current_user
 from app.schemas import VerificationResponse
+from app.services import VerificationService
+from app.models import User
 
 router = APIRouter()
 
-@router.get('/verify/{token}', response_model = VerificationResponse, status_code = status.HTTP_200_OK, summary = 'Verification link', description = 'The email will contain a link to confirm your email. This link contains a token that is required to confirm your user account. You must be a registered user with a jwt token before you can verify your email.')
-async def verify_user(request: Request, token: str, service = Depends(get_verification_service), user = Depends(get_current_user)):
+@router.get('/verify/{token}', response_model = VerificationResponse, status_code = status.HTTP_200_OK, summary = 'Verification link', description = 'The email will contain a link to confirm your email. This link contains a token that is required to confirm your user account. You must be a registered user with a jwt token before you can verify your email. Requires a jwt token.')
+async def verify_user(request: Request, token: str, service: VerificationService = Depends(get_verification_service), current_user: User = Depends(get_current_user)):
     return await service.verify_email(token)