+++ app/schemas/__init__.py
@@ -1,4 +1,5 @@
-from app.schemas.auth import UserRegister, UserLogin, UserResponse, TokenResponse, UserUpdate, UpdatePassword
+from app.schemas.auth import UserRegister, UserLogin, TokenResponse
 from app.schemas.verification import VerificationResponse
+from app.schemas.user import UserUpdate, UserResponse, UpdatePassword
 
 __all__ = ['UserRegister', 'UserLogin', 'UserResponse', 'TokenResponse', 'UserUpdate', 'UpdatePassword', 'VerificationResponse']